Send to Printer fixes
Final fixes for all routines that send to printer. They were configured inconsistently (legacy code) so thorough testing was done on all to evaluate results.

I did not use the opportunity to add the missing Print Routine for the Stat Comparison feature but would now be capable of doing so.